fix bug in DestinationStops test
By calling AllStops before setting the options in `conn_with_route`,
they weren't assigned in time for AllStops to use them to get the line
properly. This came up during some API testing where the order of stops
changed when a direction_id isn't passed (unusual), but not when a
direction_id is passed (normal behavior).