Correctly exclude stops on the Green Line.
Inverts the previous logic. Instead of checking which lines we aren't
on and unioning together their stops, we now check all the lines we
might be on and take the difference of their stops with all the GL
stops. This handles cases such as going from Boston College
(definitely on the B line) to Park Street (on the B line, but also C,
D, and E). Before this, Park Street was filtered because it is on the
other lines.