Update the event calendar filters to return the previous or next month
Why:
Currently the filters are incrementing, or subtracting, the date range
by 30 days, based on the current date. This PR updates the filters so
pressing forward or backwards filters events based on the beginning and
end date of the next or previous month.

Extras:
There is only one parameter exposed to the user (month) rather than
allowing the user to enter a custom start and end date. In the case
where an invalid date or parameter is provided, the calendar shows
upcoming events.