don't use date when linking to line information
Prevents an issue where selecting a date close to the end of the current
rating shows "No Service" because we pick a date towards the end of the
rating.

We can't go back in time either, because we could be at the beginning of the
rating and have the date in the past be before the rating starts.