use date when fetching the schedule for a trip
The API return a full datetime when getting a schedule, even when the times
would be the same on different days.  On the website side, we cache those
fetched schedules for 24 hours.  However, if a schedule for a particular trip
is fetched for today, we could still be using that cached trip tomorrow, when
it would have the wrong date associated with it.