Creates a task to import existing meetings to the new CMS
Why?
Meetings are currently being entered in a separate CMS. To keep the beta
site up-to-date with the current site, we need a way to import existing
meetings into the new CMs. This task will be in use until everyone
switches to using the new CMS.

This PR:
Creates a mix task that migrates the existing meetings to Drupal Events.
The task requires individual json files that contain the meeting
information.