Support event data migrated from the existing site
Why:
Migrated events (known as meetings in the current site) use one field,
imported_address, to show the meeting location. While the new CMS is
supporting the new Event format and the former meeting format, the view
needs to be able to display the event location or the meeting location.

PR:
Displays the meeting location unless an Event location is provided.