Adjust the meeting migration task to accept more date and time formats
Why:
The meeting data contains additional date and time range formats that
are not currently handled.

This PR:
To avoid updating the various date and time range formats in the CMS
manually, the meeting migration task now handles the additional formats
that were encountered when migrating the ~1500 records (locally).

This PR also namespaces all related files under CmsMigration, validates
that an Event start time is present, and validates the start time is
less than the end time.