fix intermittent test failure
Why:
This test currently fails between the hours of 8pm - 12am. This test
fails during those hours because we're comparing `schedule_for_trip/1`,
which uses a default value of `Timex.today` for the date (which returns
the current date based on UTC) to `schedule_for_trip/2`, which is
explicitly providing the date, which is using the current date for
America/New_York timezone.