Introduce CMS content rewriting infrastructure and use for tables
This commit introduces functionality relating to rewriting CMS body
content. It lives in the "site" app rather than "content" because in
the future it will be used to allow content writers to tap into site
functions and helpers like the circle svg mode icon helper.

The commit also includes the first use of the rewriting functionality
to make responsive tables, as a motiviating example.