Prevent the HTML scrubber from removing the mailto URI scheme in links
Why:
Currently, HtmlSanitizeEx.html5 removes the mailto URI scheme, leaving
an email address that appears to be a link but is an empty anchor tag.

This PR:
A PR has been issued to the HtmlSanitizeEx library to include mailto in
the valid schemes. This custom scrubber can be removed once that PR is
(hopefully) merged.