Notify dev team when the meeting migration task errors
Why:
The meeting migration task will be scheduled to run every so often and
otherwise forgotten about. If this task should fail, we want to know
about it.

This PR:
Sends an email to the dev team when the meeting migration task stops
due to a meeting that failed to migrate.