Prevent white underline from appearing on Event calendar nav icons
Why:
An underline text decoration is being applied to all anchor tags on
hover. This was causing an underline to appear under the fa icon
that is used to navigate months on the Event calendar page. The
underline was only appearing on the first icon as the first icon as a
tiny sliver of extra padding, allowing the underscore to appear. The
second icon does not have the same extra padding.

This PR:
Removes the underline from appearing on both Event calendar icons.