Updates links for recent news
Why:
The Recent News section links to specific News Entries by using the path
/node/:id (node was used because the URL Alias is not returned in the
recent-news payload). When creating the /news route, these links were
not updated along with the changes, which is causing the page to 404
when attempting to visit `node/:id`.

This PR:
Updates the Recent News section to link to news entries using the
news_entry_path