Updates news migration task to set the teaser field
Why:
We would like the teaser field, in the CMS, to be a required field. This
means that the teaser field must be populated when creating records
during the migration task. The data being migrated does not include a
teaser, so this tasks constructs a teaser from the news body.

This PR:
- handles an additional date format
- sets the teaser field when migrating records
- relocates the News.Blurb module to Content.Blurb
- extends the Blurb module to accept a length and default message