redirect retired or relative news and meeting paths to the correct path
Why:
When importing news and meetings from the old MBTA site, the body
sometimes includes relative links to public_meetings and news_events.
Rather than prepend these relative URLs with the old MBTA host, we would
like to support these URLs and direct the user to the new site's News
or Events page. Once the old MBTA site directs all traffic to the new
site, old URLs will still work with the new site. If the URL includes
an id, the user is directed to the corresponding record.

For example:
/about_the_mbta/public_meetings/ => /events
/about_the_mbta/public_meetings/?id=64 => /events/<find_by_migration_id>