Corrects the vertical alignment of the title on the event page
Why:
Reviewing the Boostrap docs for Card, I decided to apply the Boostrap
class directory to the header element. However, using a header element
adds the margin-bottom: .5em, which throws off the vertical alignment.

In our current version of Boostrap, Bootstrap v4.0.0-alpha.2, the
card-header class does not address the additional margin that is added
by the header element. This is address in  v4.0.0-alpha5, where the
card-header margin-bottom is set to zero.

This PR:
Adds the Boostrap class directly to the header element and updates the
card-header to set the margin-bottom to zero as alpha 5 does.