Creates one test helper to parse json files
Why:
We currently parse a json file in a number of ways in our tests. This
creates one test helper that can be used consistently to parse a json
file in /fixtures or in /priv.

Additional Notes:
Given Content.Static does not live in the test directory, Content.Static
is not able to use JsonHelpers unless the test/support directory is
compiled in all environments (it's currently only compiled for the test
environment).