Provide test coverage for the stubbed HTTPClient
Why:
The stubbed HTTPClient (Content.Static) that we use in tests is covered
by CodeCov. Most of the stubbed paths are covered by tests in the Content
app but some paths are only exercised by the Site app, which leads to
CodeCov stating that test coverage is missing. Since other apps rely on
Content.Static's interface, it feels reasonable to explicitly test the
stubbed paths.

This PR explicitly tests the paths that return a canned response.