Allow content admins to specify the table cell vertical alignment
Why:
We currently set all table data to vertical-align top but Content admins
have expressed an interest in being able to control the vertical
alignment.

Through the CMS UI, the user can specify the vertical alignment, which
adds inline style information, overriding the CSS. However, our HTML
sanitizer, HTMLSanitizeEx, is removing the word "middle" as it's not
included in the list of allowed keywords.

This PR:
- Removes the vertical alignment applied to all table cells.
- Relies on browsers using vertical-alignment: middle as the default until
  HTMLSanitizeEx merges our issued PR.