Acknowledge deleted records when importing data from Ektron
Why:
When importing data from Ektron the the CMS, we currently convert the
Ektron record to a JSON file, which is then used by another task to
import the data to the CMS. Since the JSON files are never deleted, only
updated, we're migrating records that may have been deleted on Ektron.

This PR:
Clears the specified directory of all json files, before each import, to
ensure we're not migrating stale data.