Update the content factory to use a more appropriate fixtured response
Why:
When the person factory was introduced, the CMS API only returned the
person record within the people grid paragraph. This is no longer true
as the API now includes the endpoint api/people.

This PR:
Updates the content person factory to use the stubbed json response to
api/people. This revised approach is more straight forward and saves us
unnecessary work of parsing a person record from the stubbed paragraphs
response.