Assert breadcrumbs appear in the rendered HTML
Why:
We currently test that a controller assigned the expected breadcrumbs by
checking the assigns key in the connection. If the logic that renders
breadcrumbs were removed, these tests would still pass. Testing
breadcrumbs also requires an understanding of how breadcrumbs are
structured when assigned to a connection.

This PR:
Creates a test helper to assert that the rendered html contains the
expected breadcrumb values.