avoid parsing the URL when routing in prod mode
In dev mode, we do the different dispatching to handle websockets. Since we
don't use websockets in production (yet), we can skip that parsing step and
avoid the URL decoding errors.

When we do want to use Websockets, we'll need to either fix Cowboy, or find a
way to fail better when the URL doesn't decode.