Actually, remove number_of_recent_news_suggestions for home
The home page uses 4 news suggestsions as well, but from a different
endpoint. It just pulls from the regular /news endpoint, and limits
to 4. The question of "how many?" seems a concern of the home page
so I just leave that value in the controller.

On the other hand, the /recent-news endpoint actually only returns
4 items, as configured in the CMS. So this is documented with a
module attribute in NewsEntry, and exposed via a function.