refactor Green Line predictions to do less work
Previously, we called Site.ScheduleV2Controller.Predictions.call/2 5 times: once for
each of the branches of the Green Line, plus once for the whole line.  call/2
makes a call for vehicle predictions and predictions, so we were duplicating
most of that work.

Now, we call individual functions (predictions/2 and vehicle_predictions/2)
which means we only make the remote calls we're using. This appears to knock
~5 seconds off the ScheduleV2Controller.Green tests for me, as well as having
fewer timeouts.