prevent duplicate accessible routes (#1764)
The `Site.TripPlan.Merge.merge/3` function sometimes produces duplicates
when the heads of the accessible and unknown routes don't match. For
example, with the arguments `[1, 2, 3]`, and `[2, 3]`.

Given these arguments and `needed_accessible: 2, needed_unknown: 2`, the
function does the following:

1 - compares `1` and `2`, finds they don't match, adds `1` to the result
2 - sees that `total` is `2` and `accessible` is `2` in `state`
3 - takes 1 element from unknown and 2 elements from accessible and
    appends them to the result
4 - ends up with the result `[1, 2, 2, 3]` -- which contains a duplicate

This PR uses the `merge/3` function again when directly merging two
lists to remove duplicates.