selected triangle color change to match link (#1759)
In the trip list on schedule view, the rows appear light blue when
hovered over. However, there's a triangle on the right side that stays
dark blue even when the row is light blue.

To fix this, create a light blue triangle the same shape as the dark
blue triangle and update the CSS to use that triangle when the mouse
is over the row.