fix async_assigns crashes
Using the functions `async_assign` and `await_assign_all` causes the app
to crash when the CMS takes a while to load and the await times out.

To fix this, introduce `async_assign_default` and
`await_assign_all_default`. These functions are similar except they
allow a default to be set and used when there is a time out.

The implementations are based on the implementations in `Plug.Conn` and
`Task`.