Expand lines with expanded param (#1843)
Add an `?expanded` query parameter to the `/line` and `/schedule` pages.
If that parameter is equal to one of the branches on the page, then that
branch will be expanded. This will help us have the same user experience
for users who disable JavaScript.

Change the expand/collapse branch button to a link. This allows users
who have disabled JavaScript or who use a browser that doesn't support
JavaScript to navigate the website.

This functionality used to exist in the application, but was removed in
[Stop bubbles master](https://github.com/mbta/dotcom/pull/1524/files).
This is largely a re-implementation of that.