Ensure path aliases from Drupal containing slashes do not result in encoded URLs (#1949)
* Ensure path aliases from Drupal containing slashes do not result in encoded URLs

The Phoenix path helpers do not expect slashes in the resource name. If
a resource with a slash is provided, the helper will return an encoded
URI. We need to decode this URI to create proper links on the website.

Note that part of this work was done on the initial PR. This PR covers
the cases that were missed.