fix dialyzer confusion
It looks like Dialyzer is confused about the generated function
`start_link/0` referencing the `opts` that were inserted into the module as
`@opts`. To allow Dialyzer to see the right types, we insert the opts as a
function `opts/0`, and update `start_link/0` and `default_ttl/0` to use that
instead.