RepoCache: re-check the cache once the write lock is acquired
If the function to write the cache is slow, and many requests for the lock
queue up behind, they'll all end up writing the cache (or timing out and
crashing). This change adds a re-check of the cache once the write lock is
acquired: if the cache has already been written, we don't bother calling the
function.