Pass holidays down to service selector
In preparation for using knowledge of when upcoming holidays are to
correctly call out holiday schedules in the schedule finder's service
selector, here we arrange to pass the holidays down the component tree
to the ServiceSelector. We also differentiate between the "date" and the
"formatted_date", the former being kept in a more easily parseable format.