Upcoming/Recent Departures
Switch from using the predictions_api for upcoming departures to the journey-based response in the new finder_api. This keeps the schedule data in the same general shape as those listed in the trip list below upcoming departures, making it easier to adopt a shared accordion UI for both sections. Journeys here are "enhanced" to include more detailed prediction values for use in the front-end.

Upcoming Departures now includes recently departed trips for the selected origin, but only for South Station, North Station, and Back Bay, as these are the only stops that are tagged with a "Departed" status. Furthermore, as long as this "Departed" prediction is retained, we can display it in Upcoming Departures, up to the retention policy time which has been set to 10 minutes. Various routes have different internal retention times for departure predictions, and cannot be relied on 100% to always be present for every trip.